requirements.txt
# Core
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Deep learning
torch>=1.10.0
torchvision>=0.11.0
Pillow>=8.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Dimensionality reduction
umap-learn>=0.5.3

# Metrics & evaluation
lpips>=0.1.4      # Optional, only if LPIPS metric needed
scikit-image>=0.18.0

# Detection evaluation (COCO)
pycocotools>=2.0.6

# Utilities
tqdm>=4.62.0
ujson>=5.0.0

# Optional (for TSNE)
openTSNE>=0.6.0


# klpips is optional — install it only if you want LPIPS metric (--lpips flag).
# ujson is also optional — but helps speed up COCO JSON loading.
# openTSNE is only needed if you want to use a faster t-SNE than sklearn’s.

# pip install -r requirements.txt